Re-introduced "How to Reach Me" section
Re-introduced the "How to Reach Me" section. It's mostly a duplicate of the links at the top of my page, but it adds a bit more detail as to what each one is for. I did this because the Discord tag link is caps-sensitive, and it doesn't reflect the correct capitalization in the link at the top.